Move gatsby-node logic to .esm.js and reference
